Fix lint issues and improve code organization
- Extract utility functions from Leaderboard component to separate utils/leaderboard.ts file
- Remove buttonVariants export from shadcn Button component to prevent fast-refresh warnings
- Remove unused type imports
- Improve code maintainability by separating concerns
